chore: move code from internal repo (#11)
Moving code as-is from the current internal repo. The repo was updated few minutes ago, so it should be the latest code version.

The code was not edited, but there was some re-formatting happening, probably due to a different prettier settings. I have copied prettier params from our package as well, but it does not seem working...