feat(dashboard): use INHERIT period by default (#108)
We should set INHERIT by default to ensure that every graph has the preferred time resolution instead of the AUTO one (which can be 5m or 1h based on the range). It is better to have it set like this, since it will better correspond to alarms.

---

_By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license_

(cherry picked from commit b6b28a462ec4ea5e8bbabcbcfe3353463b644d46)