fix(UsageAlarmFactory): allowing addition of multiple alarms from the same class (#279)
Fixes # Addition of multiple alarms from the sameclass

UsageAlarmFactory is creating same AlarmSuffix  for every alarm that is being added in a class which leads to creation of duplicate alarms, as they have same name. This fix appends the UsageType to the alarm name thus allowing users to create alarm on avg, p50,p90... all metrics from the same class.

---

_By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license_