feat(ec2): Take into account EBS for disk metrics
EC2 provides different metrics for Nitro-based instances. These metrics have the
prefix `EBS` instead of `Disk`.

The metrics are either `Disk*` or `EBS*`, never both at the same time for an
instance.

In the case of an AutoScalingGroup, then we'll have the 2 metrics if the EC2
instances within that group are mixed (both nitro and classic), so in this case
averaging both metrics make sense.