feat(ec2): Also plot EBS disk metrics when NITRO instance type is specified (#321)
EC2 provides different metrics for Nitro-based instances. These metrics have the prefix `EBS`.

These instances either have the `Disk*` or `EBSDisk*` metrics, never both. Because of this I thought in using [metric math](https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/using-metric-math.html) where if any of the time series is empty, it would be removed, and then it would display a single metric. This gets trickier when monitoring an ASG because it can have both metrics there, so what should be done with the 2 time series? An average of them? If you think this would be a better idea, please advise so I change the implementation.

---

_By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license_