feat(elbv2): Accept imported load balancer and target groups (#329)
After CDK 2.65.0, it is  now possible to obtain metrics through the interfaces. This opens up the possibility of creating dashboards for the imported versions of these resources.

PS: The tests have such large diffs because of the extra indentation.  The code is very similar, I just added the extra factory (`lbResourcesFactory`), but git is confused here.

---

_By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license_