feat(dashboards): improve customizability (#349)
# Change Info
This change allows for improved customization of dashboards. It will allow consumers of this package to control how many dashboards they generate, the configuration of those dashboards, and the widgets on those dashboards. This probably helps address this related github issue: https://github.com/cdklabs/cdk-monitoring-constructs/issues/66. 

This change is backwards compatible and consumers must make explicit changes to their code if they would like to take advantage of the additional dynamism around dashboard generation this contribution provides. As a consequence of this, the dashboard configuration provided by this contribution is not as robust or as flexible as I would like it to be, but this comes at the trade-off of being minimally intrusive to the existing codebase. 

# Testing Done
I've used my own CDK application to test this and have verified that these new changes allow me to create a custom number of dashboards according to my own configuration. 

# Backwards Compatibility
This change was made backwards compatible by retrofitting the existing `DefaultDashboardFactory` to also be an implementation of `IDynamicDashboardFactory` with an opinionated implementation which creates three dashboards: Summary, Details, and Alarms. 

## Backwards Compatibility Testing Done
I also tested this for backwards compatibility by: 

1. Using my own CDK application by using the existing constructs exposed by the `MonitoringFacade` and verified that `monitoring.addSegment` and `monitoring.monitorXYZ` calls populated the default dashboards as expected
2. Verifying that the jest snapshots for the `DefaultDashboardFactory` did not change. 

# How to consume / Documentation
I've added a README section on how to consume these changes. 

# Limitations of this approach
1. This does not really support minor-changes to existing dashboard content. If you want to use this, it amounts to a full swap & replace. Re-arrangement of existing widgets to new dashboards can easily be done, but small modifications to existing widgets are not cheap with this change. 
3. This change introduces some coupling between dashboard types and IDynamicDashboardSegments: when creating IDynamicDashboardSegment implementations, you are required to know which dashboards you are generating. 

However, since this contribution is minimally intrusive, I thought the above were decent trade-offs to form the basis of PR discussions. 

# Forward Thinking / Additional Augmentations
Organizations can use this to configure their own defaults for service teams in their organization and setup best practices which can easily be consumed by service teams when creating their dashboards. Further, such organizations can even introduce additional dashboard types (e.g. for higher level operational or business review) which can exist separately from service/engineering dashboards which are intended to have higher detail and be consumed by more technical audiences.

---

_By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license_