feat(alarm): add minSampleCountToEvaluateDatapoint to CustomMonitoring (#458)
Fixes #452

Follow up to
https://github.com/cdklabs/cdk-monitoring-constructs/pull/453:

* (feat) Exposing the new `minSampleCountToEvaluateDatapoint` through
CustomMonitoring
* (fix) Fixing the `minSampleCountToEvaluateDatapoint` MathExpression's
period as it defaults to 5 minutes. This didn't come up during testing
as I tested it using 5 minute period. Apparently, if we don't set the
period on MathExpression explicitly, it overrides all child metrics to 5
minute,
[reference](https://github.com/aws/aws-cdk/blob/db21fefc2dc76eb4ff306fa41652ab6a6cc95e42/packages/aws-cdk-lib/aws-cloudwatch/lib/metric.ts#L606).
To avoid similar situations in the future, extended the unit test to
cover custom periods.

---

_By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license_