feat: add fluentbit monitoring (#470)
This adds FluentBit monitoring. FluentBit is default technology used by Firelens, which is suggested log router to use with ECS. Current implementation doesn't support any alarms, but can be extended if there will be interest. 

Photo of this dashboard segment from our team code that this change is based upon.
![screenshot_2023-12-12_at_15 48 19](https://github.com/cdklabs/cdk-monitoring-constructs/assets/28189907/35b74e77-3906-4de0-8027-e987a81b2f85)

---

_By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license_