fix(lambda): fix iterator age dedupe
Our team has tiered alarms on the max iterator age of events going to a Lambda function so that we get higher severity tickets for higher iterator ages which indicates worse backups in processing.

However, these are not currently deduping into the same ticket which I root caused to the `alarmDedupeStringSuffix` being missing on the `addIteratorMaxAgeAlarm` which the Lambda monitoring uses.

This PR fixes this.