feat(lambda): add isOffsetLag prop for alarms/dashboards for OffsetLag metric (#574)
Lambda has added support for connecting to Kafka streams via Event
Source Mappings (see:
https://docs.aws.amazon.com/lambda/latest/dg/with-msk.html). This comes
with the addition of a new CW metric under the AWS/Lambda namespace,
`OffsetLag` (see:
https://aws.amazon.com/blogs/compute/offset-lag-metric-for-amazon-msk-as-an-event-source-for-lambda/).
This metric is different than the existing IteratorAge metric used for
Kinesis streams. OffsetLag represents the difference between the last
record written to a Kafka topic and the last record that the function's
consumer group has processed (note the metric is in # of records, it's
not a time metric). See here:
https://docs.aws.amazon.com/lambda/latest/dg/monitoring-metrics.html.

Defaulted the monitor to `false` since it's a new prop and is not widely
used.

---

_By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license_

Co-authored-by: Connor Koch <connor@Connors-MBP.attlocal.net>