feat(dashboard): add support for dashboard variables (#639)
Fixes https://github.com/cdklabs/cdk-monitoring-constructs/issues/391

Not sure if this change is really needed — I’ve been creating dashboards with variables by just grabbing the underlying dashboards and attaching variables that way. That said, it’s a nice cleanup and does make the code a bit cleaner. Totally fine to close the PR if you don’t think it’s worth it.

---

_By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license_