build!: upgrade to JSII@5
This removes some APIs/types that had replacements added previously in order to avoid using
generics and tuples, which JSII@5 does not support.

The BaseMetricFactory might affect other extensions of that abstract class outside of this library,
which would need to be confirmed by reviewers.

This does not address Rosetta doc errors, which are still disabled.