feat: introduce replacement types for aspect (#687)
This is extracted from #686 and an incremental step towards upgrading to
JSII@5.x.

This introduces new types to replace the usage of the generic type that
needs to be removed since JSII@5.x doesn't support generics as part of
constructs' APIs.

This does kind of spam new interfaces into the API, but I think that's
an acceptable trade-off (the alternative is just removing this aspect,
or moving it to a TS-only library).

`MonitoringAspectType` would be removed as part of a later breaking
change.

---

_By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache-2.0 license_

---------

Co-authored-by: arkon <arkon@users.noreply.github.com>
Co-authored-by: Caymon / Kayo <24500457+giantcow@users.noreply.github.com>